# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For DH Mobile ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/variant-attribute-values": {
            "get": {
                "tags": [
                    "VariantAttributeValues"
                ],
                "summary": "Lấy danh sách liên kết biến thể - giá trị thuộc tính",
                "operationId": "9ea7bb8fe666ae49af9a38e5a247e550",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            },
            "post": {
                "tags": [
                    "VariantAttributeValues"
                ],
                "summary": "Tạo liên kết biến thể - giá trị thuộc tính mới",
                "operationId": "68e3288822140b408e8a9438bfe8d72e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "variant_id",
                                    "value_id"
                                ],
                                "properties": {
                                    "variant_id": {
                                        "type": "integer"
                                    },
                                    "value_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    }
                }
            }
        },
        "/api/variant-attribute-values/{id}": {
            "get": {
                "tags": [
                    "VariantAttributeValues"
                ],
                "summary": "Lấy chi tiết liên kết biến thể - giá trị thuộc tính",
                "operationId": "48ac7a8b6db1d7344d9022746e0e81f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "put": {
                "tags": [
                    "VariantAttributeValues"
                ],
                "summary": "Cập nhật liên kết biến thể - giá trị thuộc tính",
                "operationId": "acfcc3ca74daf2f7cf9a5b8464bedfa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "variant_id": {
                                        "type": "integer"
                                    },
                                    "value_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "delete": {
                "tags": [
                    "VariantAttributeValues"
                ],
                "summary": "Xóa mềm liên kết biến thể - giá trị thuộc tính",
                "operationId": "ee6c33361cc8d36f4458c7f288e7a121",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
        "/api/variant-attribute-values/{id}/restore": {
            "put": {
                "tags": [
                    "VariantAttributeValues"
                ],
                "summary": "Khôi phục liên kết đã xóa mềm",
                "operationId": "86472b2e29f4e55bec6f6d8c6e7df507",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Khôi phục thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
        "/api/variant-attribute-values/{id}/force": {
            "delete": {
                "tags": [
                    "VariantAttributeValues"
                ],
                "summary": "Xóa vĩnh viễn liên kết đã xóa mềm",
                "operationId": "1be28d68037c9a5c16d56d85f049f9cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa vĩnh viễn thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}